+++ app/Http/Controllers/Cms/User/User/UsersController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\Cms\User\User;
 use App\Http\Controllers\Cms\User\User\Requests\StoreUserRequest;
 use App\Http\Controllers\Cms\User\User\Requests\UpdateUserRequest;
 use App\Models\User\User;
+use App\Policies\Abilities;
 use App\Services\Cms\User\GroupsService;
 use App\Services\Cms\User\UsersService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Controllers\Controller;
@@ -35,9 +37,12 @@ class UsersController extends Controller
      * Display a listing of the resource.
      *
      * @return Factory|View
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize(Abilities::VIEW_ANY, User::class);
+
         return view('cms.user.index', [
             'users' => $this->usersService->paginationList(),
         ]);
@@ -47,9 +52,12 @@ class UsersController extends Controller
      * Show the form for creating a new resource.
      *
      * @return Factory|View
+     * @throws AuthorizationException
      */
     public function create()
     {
+        $this->authorize(Abilities::CREATE, User::class);
+
         return  view(
             'cms.user.create',
             [
@@ -63,9 +71,12 @@ class UsersController extends Controller
      *
      * @param  StoreUserRequest  $request
      * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
      */
     public function store(StoreUserRequest $request)
     {
+        $this->authorize(Abilities::CREATE, User::class);
+
         $url = $this->usersService->store($request);
 
         return redirect($url);
@@ -76,9 +87,12 @@ class UsersController extends Controller
      *
      * @param User $user
      * @return Factory|View
+     * @throws AuthorizationException
      */
     public function show(User $user)
     {
+        $this->authorize(Abilities::VIEW, $user);
+
         return view(
             'cms.user.show',
             [
@@ -93,9 +107,12 @@ class UsersController extends Controller
      *
      * @param User $user
      * @return Factory|View
+     * @throws AuthorizationException
      */
     public function edit(User $user)
     {
+        $this->authorize(Abilities::UPDATE, $user);
+
         return view(
             'cms.user.edit',
             [
@@ -112,9 +129,12 @@ class UsersController extends Controller
      * @param UpdateUserRequest $request
      * @param User $user
      * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $this->authorize(Abilities::UPDATE, $user);
+
         $url = $this->usersService->update($user, $request);
 
         return redirect($url);
@@ -125,9 +145,12 @@ class UsersController extends Controller
      *
      * @param User $user
      * @return RedirectResponse|Redirector
+     * @throws AuthorizationException
      */
     public function destroy(User $user)
     {
+        $this->authorize(Abilities::DELETE, $user);
+
         $url = $this->usersService->destroy($user);
 
         return redirect($url);
